refactor(observability): share claude usage normalization
