+++ src/infra/claude-headless/stream-json-lines.ts
@@ -1,14 +1,10 @@
-import { USAGE_MISSING_REASONS } from '../../core/logging/contracts.js';
 import type { ProviderUsageSnapshot } from '../../core/models/response.js';
+import { extractClaudeProviderUsage } from '../claude/usage.js';
 
 function toRecord(value: unknown): Record<string, unknown> | undefined {
   return value !== null && typeof value === 'object' && !Array.isArray(value)
     ? (value as Record<string, unknown>)
     : undefined;
-}
-
-function toNumber(value: unknown): number | undefined {
-  return typeof value === 'number' && Number.isFinite(value) ? value : undefined;
 }
 
 function pickString(source: Record<string, unknown> | undefined, keys: string[]): string | undefined {
@@ -171,44 +167,6 @@ function extractStructuredOutput(event: Record<string, unknown>): Record<string,
   return toRecord(structuredOutput);
 }
 
-function extractProviderUsage(event: Record<string, unknown>): ProviderUsageSnapshot | undefined {
-  const usage = toRecord(event.usage);
-  if (!usage) {
-    return undefined;
-  }
-
-  const inputTokens = toNumber(usage.input_tokens);
-  const outputTokens = toNumber(usage.output_tokens);
-  const cacheCreationInputTokens = toNumber(usage.cache_creation_input_tokens);
-  const cacheReadInputTokens = toNumber(usage.cache_read_input_tokens);
-  if (inputTokens === undefined || outputTokens === undefined) {
-    return {
-      usageMissing: true,
-      reason: USAGE_MISSING_REASONS.TOKENS_MISSING,
-    };
-  }
-
-  const providerUsage: ProviderUsageSnapshot = {
-    inputTokens,
-    outputTokens,
-    totalTokens: inputTokens + outputTokens,
-    usageMissing: false,
-  };
-  const cachedInputTokens = cacheCreationInputTokens !== undefined && cacheReadInputTokens !== undefined
-    ? cacheCreationInputTokens + cacheReadInputTokens
-    : cacheReadInputTokens ?? cacheCreationInputTokens;
-  if (cachedInputTokens !== undefined) {
-    providerUsage.cachedInputTokens = cachedInputTokens;
-  }
-  if (cacheCreationInputTokens !== undefined) {
-    providerUsage.cacheCreationInputTokens = cacheCreationInputTokens;
-  }
-  if (cacheReadInputTokens !== undefined) {
-    providerUsage.cacheReadInputTokens = cacheReadInputTokens;
-  }
-  return providerUsage;
-}
-
 function isSuccessfulResultEvent(event: Record<string, unknown>): boolean {
   if (event.is_error === true || event.isError === true) {
     return false;
@@ -258,7 +216,7 @@ export function aggregateResultFromStdout(stdout: string): StreamJsonStdoutResul
     success = isSuccessfulResultEvent(root);
     error = success ? undefined : extractResultError(root, resultContent);
     structuredOutput = extractStructuredOutput(root);
-    providerUsage = extractProviderUsage(root);
+    providerUsage = extractClaudeProviderUsage(root.usage);
   }
 
   const normalizedDisplayText = displayText.trim();
